% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.cell.from.subcells.R
\name{estimate.cell.from.subcells}
\alias{estimate.cell.from.subcells}
\title{Richness estimate for a cell based on cells from a smaller grid within the one of interest}
\usage{
estimate.cell.from.subcells(focal.cell,
  data.long.format.grid.subgrid.species, indice.estimate = "chao",
  colname_cell, colname_subcell, colname_species)
}
\arguments{
\item{focal.cell}{id of the focal cell for which richness estimates are needed.}

\item{data.long.format.grid.subgrid.species}{a data.frame in long format (1 row = 1 occurence) with species, cell id for a grid (e.g. 50x50km) and cell id for a smaller grid (e.g. 10x10km).}

\item{indice.estimate}{character, one of the index available in specpool: chao (default), jack1, jack2, boot.}

\item{colname_cell}{character, the name of the column that contains the cell id of the largest grid.}

\item{colname_subcell}{character, the name of the column that contains the cell id of the small grid that will be used as sampling unit.}

\item{colname_species}{character, the name of the column that contains the species.}
}
\description{
This function computes richness estimates for a focal cell and returns the percentage of observed richness compared to one estimate index. To estimate richness, sampling units used are cells from a smaller grid that are included in the focal cell.
}
\examples{
estimate.cell.from.subcells()


}
\keyword{completeness}
