% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/focal.cell.beta.mean.R
\name{focal.cell.beta.mean}
\alias{focal.cell.beta.mean}
\title{Mean beta diversity for focal cell}
\usage{
focal.cell.beta.mean(focal.cell, grid.with.species, radius = 10000,
  X.long, Y.lat, index.family = "jac", component = "dis")
}
\arguments{
\item{focal.cell}{rownames of the focal cell to estimate completeness.}

\item{grid.with.species}{a data.frame with long, lat, id.cell and species occurence data.}

\item{radius}{will define the number of neighbor cells. Default to 10000 (1 cell for 10x10km grid).}

\item{X.long}{character, the name of the column that contains X or longitude.}

\item{Y.lat}{character, the name of the column that contains Y or latitude.}

\item{index.family}{family of dissimilarity indices, partial match of "sorensen" or "jaccard".}

\item{component}{which component (turnover, nestedness, total) of the beta-diversity is returned by the function: "sim", "sne" or "sor", or jtu", "jne" or "jac".}

\item{grid.with.species}{a data.frame with long, lat, id.cell and species occurence data.}
}
\description{
This function computes the mean beta-diversity between the focal cell and its neighboring cells (defined according to the radius).
}
\examples{
focal.cell.beta.mean()
}
\keyword{completeness}
