% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/focal.cell.completeness.proxy.R
\name{focal.cell.completeness.proxy}
\alias{focal.cell.completeness.proxy}
\title{Proxy of completeness}
\usage{
focal.cell.completeness.proxy(focal.cell, grid.with.species,
  radius = 10000, focal.cell.in.estimate = T, X.long, Y.lat, index)
}
\arguments{
\item{focal.cell}{rownames of the focal cell to estimate completeness.}

\item{grid.with.species}{a data.frame with long, lat, id.cell and species occurence data.}

\item{radius}{will define the number of neighbor cells. Default to 10000 (1 cell for 10x10km grid).}

\item{focal.cell.in.estimate}{does the focal need to be taken into account to estimate diversity. default to TRUE}

\item{X.long}{character, the name of the column that contains X or longitude.}

\item{Y.lat}{character, the name of the column that contains Y or latitude.}

\item{index}{character, one of the index available in specpool: chao, jack1, jack2, boot.}

\item{grid.with.species}{a data.frame with long, lat, id.cell and species occurence data.}
}
\description{
This function compares the richness of one focal cell to the richness estimated from the neighboring cells with one the following indices: Chao, jackknife (1 or 2), bootstrap.
}
\examples{
focal.cell.completeness.proxy()
}
\keyword{completeness}
